+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # QuickPrint Changelog
 
 
+## v1.0.2: 2022-05-28
+
+* Added `advanceXCoarse()`, which provides basic "snap to grid"-like positioning of the cursor X position.
+* Added `setXMin()`, which moves the cursor X only if the current X position is less than the requested position.
+* Split the string-accepting logic of `advanceX()` into a separate function: `advanceXStr()`.
+* Added some single-axis versions of cursor position methods:
+  * `setPosition()`: `setXPosition()` and `setYPosition()`
+  * `getPosition()`: `getXPosition()` and `getYPosition()`
+  * `movePosition()`: `moveXPosition()` and `moveYPosition()`
+
+
 ## v1.0.1: 2022-05-16
 
 * Started changelog.